TesMenu Full

+++ src/main/java/am/aca/quiz/software/repository/TestRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.GetMapping;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
@@ -13,6 +14,6 @@ public interface TestRepository extends JpaRepository<TestEntity,Long> {
     TestEntity findTestEntitiesByTest(String test);
 
     @Query(value = " SELECT t.id FROM test as t INNER JOIN question_test as qt ON t.id=qt.test_id INNER JOIN questions as q ON qt.question_id=q.id WHERE q.topic_entity_id=?1",nativeQuery = true)
-    List<Long> findTestByTopicId(Long id);
+    List<BigInteger> findTestByTopicId(Long id);
 
 }
