+++ src/main/java/am/aca/quiz/software/service/implementations/TestServiceImp.java
@@ -8,6 +8,7 @@ import am.aca.quiz.software.service.interfaces.TestService;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PatchMapping;
 
+import java.math.BigInteger;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
@@ -61,7 +62,7 @@ public class TestServiceImp implements TestService {
         return testRepository.findById(id).get();
     }
 
-    public List<Long> findTestIdByTopicId(Long id){
+    public List<BigInteger> findTestIdByTopicId(Long id){
         return testRepository.findTestByTopicId(id);
     }
